Move provider selector button for reuse
